app/requirements.txt
sanic==21.12.1
huey==2.4.3
PyGithub==1.55
GitPython==3.1.27
PyYAML==6.0
pymongo==4.0.2
